third_party/blink/web_tests/platform/linux-chrome/external/wpt/navigation-api/scroll-behavior/after-transition-intercept-handler-modifies-expected.txt

This is a testharness.js-based test.
[FAIL] scroll: state should be saved before intercept handlers run
  assert_equals: expected 100 but got 1100
Harness: the test ran to completion.
